requirements.txt
google-genai==0.5.0
langchain-google-genai==2.0.8
langchain==0.3.25
langgraph==0.4.8
python-dotenv==1.1.0
faiss-cpu==1.12.0
langchain-community==0.3.25
pypdf==5.6.0